requirements.txt
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cryptography==2.9.2
defusedxml==0.6.0
eight==1.0.0
Flask==1.1.2
flask-saml2==0.3.0
future==0.18.2
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
pycparser==2.20
pyOpenSSL==17.5.0
pytz==2020.1
signxml==2.8.0
six==1.15.0
Werkzeug==1.0.1
